File restructuring (one class per file)
